requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0.post1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Werkzeug==3.1.3
